Move fetchBuckets to GoogleStorage
